Don't render gif grid while gif list refreshes
Whenever a user performs a search for the first time, we render an
entirely new list of gifs. If we are rendering the grid the whole time,
there is a weird user experience where the user has performed a search
but the old gifs are still displaying while the request completes,
giving the impression that the search form didn't do anything.

By not rendering the grid while the new search request completes, the
user instead sees the gif grid disappear for a moment before reappearing
with the results of the new search.

https://github.com/Euraldius/giphy-browser/projects/1#card-19145624