Don't infinite scroll if all gifs loaded
In general to it feeling like good practice, Giphy rate limits us, so we
don't want to keep making requests for more gifs if we've already loaded
all of the existing gifs for a given search.

I'm not worried about this for trending gifs, since there are ~100000 of
those at any given time. Theoretically someone could write a script to
attack us by scrolling out of trending gifs and getting us rate limited
but…this is just a code challenge, so I'm not worried about that right
now.

Also, this commit fixes a bug in the request offsets that was
artificially increases the offset by 1. (Which I thought was necessary
because I did the math wrong.)

https://github.com/Euraldius/giphy-browser/projects/1#card-19080581