Refactor handling of state
I noticed that there was a lot of duplication between the handling of
trending gifs state and search gifs state. This duplication also
required that the App container have extra logic for determining whether
either state had parameters it needed to know (like `isFetching`).

This commit consolidates the state handling to make it simpler. For
example, there is only one list of gifs displayed at a time, so the
state now only has to maintain one list, instead of one for trending
gifs and one for searched gifs.

As a result, I expect future development on the app would be less likely
to produce bugs resulting from an unnecessarily complex state, something
with which I have struggled.

:)